fix(labels): consolidate remaining string literals and reduce file size (#2520)
- Extract LabelAnomaly and collect_label_anomalies to label_anomalies.py (350 -> 274 lines in labels.py)
- Add ORCHESTRA_GOVERNED_LABEL constant and use it in governance_utils.py
- Update lazy import chains in services/__init__.py and shared/__init__.py
- Fix deep import violation in governance_utils.py
