+++ src/vibe3/services/shared/labels.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import functools
-from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from loguru import logger
@@ -162,6 +161,8 @@ _STATE_PRIORITY_ORDER = [
     "ready",
 ]
 
+ORCHESTRA_GOVERNED_LABEL = "orchestra-governed"
+
 
 # ---------------------------------------------------------------------------
 # Predicates (pure functions, no I/O)
@@ -270,81 +271,4 @@ def classify_dispatch_eligibility(
     return reasons
 
 
-@dataclass(frozen=True)
-class LabelAnomaly:
-    """Audit finding for a single issue's label state."""
-
-    issue_number: int
-    rule: str  # roadmap_conflict | multi_state | orphan_execution | orphan_orchestra
-    removed: list[str]
-    added: list[str]
-
-
-def collect_label_anomalies(
-    labels: list[str],
-    *,
-    has_local_flow: bool,
-    is_manager_issue: bool,
-) -> list[LabelAnomaly]:
-    """Collect all label anomalies for one issue.
-
-    Returns list of anomalies (empty = no issues found).
-    Rules evaluated in priority order; roadmap_conflict suppresses
-    multi_state and orphan_execution rules.
-    """
-    removed: list[str] = []
-    added: list[str] = []
-    rules: list[str] = []
-
-    # Rule 1: roadmap + state conflict
-    r1 = (
-        [lb for lb in labels if lb.startswith("state/")]
-        if has_roadmap_label(labels)
-        else []
-    )
-    if r1:
-        removed.extend(r1)
-        rules.append("roadmap_conflict")
-
-    if not r1:
-        # Rule 2: multiple state labels
-        conflicts = get_conflicting_states(labels)
-        if conflicts:
-            removed.extend(conflicts)
-            rules.append("multi_state")
-
-        # Rule 3: orphan execution state (manager issues only)
-        if is_manager_issue and not has_local_flow and has_execution_state(labels):
-            exec_labels = [
-                lb
-                for lb in get_state_labels(labels)
-                if lb.removeprefix("state/") in _EXECUTION_STATES
-            ]
-            if exec_labels:
-                removed.extend(exec_labels)
-                added.append("state/ready")
-                rules.append("orphan_execution")
-
-    # Rule 4: orphan orchestra-governed (manager issues only)
-    if is_manager_issue and has_orchestra_governed(labels):
-        has_state = bool(get_state_labels(labels))
-        has_roadmap = has_roadmap_label(labels)
-        if not has_state and not has_roadmap:
-            removed.append("orchestra-governed")
-            rules.append("orphan_orchestra")
-
-    if not rules:
-        return []
-
-    # Deduplicate while preserving order
-    removed = list(dict.fromkeys(removed))
-    added = list(dict.fromkeys(added))
-
-    return [
-        LabelAnomaly(
-            issue_number=0,  # caller sets this
-            rule=", ".join(rules),
-            removed=removed,
-            added=added,
-        )
-    ]
+# LabelAnomaly and collect_label_anomalies moved to label_anomalies.py
